+++ internal/cli/arg_context.go
@@ -8,7 +8,7 @@ type argContext int
 // The Enum used to keep track of the argument context.
 const (
 	// The context where arguments are interpreted as input files.
-	contextInputFile argContext = iota
+	contextDefault argContext = iota
 
 	// The context where arguments are interpreted as a configuration file.
 	contextConfigFile