+++ internal/cli/parse.go
@@ -13,6 +13,8 @@ line argument for the wordrow program. This will provide a custom struct
 package cli
 
 import (
+	"fmt"
+
 	"github.com/ericcornelissen/wordrow/internal/errors"
 	"github.com/ericcornelissen/wordrow/internal/logger"
 	"github.com/ericcornelissen/wordrow/internal/strings"
@@ -23,9 +25,17 @@ func noArgumentsProvided(args []string) bool {
 	return len(args) == 1
 }
 
-// Check if a certain argument is an option.
-func argumentIsOption(arg string) bool {
-	return strings.HasPrefix(arg, "-")
+// Parse an option argument as a (set of) flag(s).
+func parseArgumentAsAlias(
+	alias string,
+	arguments *Arguments,
+) (newContext argContext, err error) {
+	for _, char := range alias[1:] {
+		option := fmt.Sprintf("-%c", char)
+		newContext, err = parseArgumentAsOption(option, arguments)
+	}
+
+	return newContext, err
 }
 
 // Parse an option argument and get the new argument context.
@@ -33,14 +43,14 @@ func parseArgumentAsOption(
 	option string,
 	arguments *Arguments,
 ) (argContext, error) {
-	newContext := contextInputFile
+	newContext := contextDefault
 	switch option {
 	case helpFlag.name:
 		arguments.help = true
 	case versionFlag.name:
 		arguments.Version = true
 
-		// Flags
+	// Flags
 	case dryRunFlag.name:
 		arguments.DryRun = true
 	case invertFlag.name, invertFlag.alias:
@@ -72,7 +82,7 @@ func parseArgumentAsValue(
 	arguments *Arguments,
 ) {
 	switch context {
-	case contextInputFile:
+	case contextDefault:
 		arguments.InputFiles = append(arguments.InputFiles, value)
 	case contextConfigFile:
 		arguments.ConfigFile = value
@@ -83,7 +93,7 @@ func parseArgumentAsValue(
 	}
 }
 
-// Parse a single argument, value or option.
+// Parse a single argument as a value or option/flag.
 //
 // The function sets the error if the argument could not be parsed (in the
 // provided context).
@@ -91,22 +101,23 @@ func doParseOneArgument(
 	arg string,
 	context argContext,
 	arguments *Arguments,
-) (argContext, error) {
-	if argumentIsOption(arg) {
-		if context != contextInputFile {
+) (newContext argContext, err error) {
+	if strings.HasPrefix(arg, "-") {
+		if context != contextDefault {
 			return context, errors.Newf("Missing value for %s option", context)
 		}
 
-		newContext, err := parseArgumentAsOption(arg, arguments)
-		if err != nil {
-			return context, err
+		if strings.HasPrefix(arg, "--") {
+			newContext, err = parseArgumentAsOption(arg, arguments)
+		} else {
+			newContext, err = parseArgumentAsAlias(arg, arguments)
 		}
-
-		return newContext, nil
+	} else {
+		parseArgumentAsValue(arg, context, arguments)
+		newContext = contextDefault
 	}
 
-	parseArgumentAsValue(arg, context, arguments)
-	return contextInputFile, nil
+	return newContext, err
 }
 
 // Parse a slice of arguments that contains at least one program argument.
@@ -116,7 +127,7 @@ func doParseOneArgument(
 func doParseProgramArguments(args []string) (Arguments, error) {
 	var arguments Arguments
 
-	context := contextInputFile
+	context := contextDefault
 	for _, arg := range args {
 		newContext, err := doParseOneArgument(arg, context, &arguments)
 		if err != nil {
@@ -126,7 +137,7 @@ func doParseProgramArguments(args []string) (Arguments, error) {
 		context = newContext
 	}
 
-	if context != contextInputFile {
+	if context != contextDefault {
 		return arguments, errors.New("More arguments expected")
 	}